Just some junk
Here's some junk, don't mind me

+++ Assets/Objects/Orb.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Orb : MonoBehaviour
 {
-    public int despawnTime;
+    public float despawnTime;
 
     public Animator anim;
